requirements.txt
chromadb~=0.6.3
python-dotenv~=1.0.1
PyYAML~=6.0.2
pydantic~=2.10.6
fastapi~=0.115.11
aiostream~=0.6.4
uvicorn~=0.34.0

llama-index-core~=0.12.20

llama-index-vector-stores-chroma
llama-index-readers-file
llama-index-readers-database
llama-index-readers-web

llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-llms-lmstudio
llama-index-llms-openai

selenium
wikipedia
duckduckgo_search